Fix _LocIndexer._setitem_with_indexer
This method is responsible for `DataFrame.loc[i] = value`. It uses
append in its inner workings.

Though I haven't tested yet, I think that setting sort to False is the
right choice here, because, never would we want to sort the DataFrame
columns when we are just adding some rows.